moved pointerup event registration to magic-side-nav to reduce number of fired events
