+++ src/app/core-ui/magic-side-nav/magic-side-nav.component.ts
@@ -10,6 +10,7 @@ import {
   OnInit,
   output,
   signal,
+  AfterViewInit,
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
@@ -24,6 +25,12 @@ import { NavMatMenuComponent } from './nav-mat-menu/nav-mat-menu.component';
 import { TaskService } from '../../features/tasks/task.service';
 import { LayoutService } from '../layout/layout.service';
 import { magicSideNavAnimations } from './magic-side-nav.animations';
+import { Subscription } from 'rxjs';
+import { ScheduleExternalDragService } from '../../features/schedule/schedule-week/schedule-external-drag.service';
+import { Log } from '../../core/log';
+import { TODAY_TAG } from '../../features/tag/tag.const';
+import { DragDropRegistry } from '@angular/cdk/drag-drop';
+import { WorkContextType } from '../../features/work-context/work-context.model';
 
 const COLLAPSED_WIDTH = 64;
 const MOBILE_NAV_WIDTH = 300;
@@ -49,13 +56,17 @@ const FOCUS_DELAY_MS = 10;
   },
   animations: magicSideNavAnimations,
 })
-export class MagicSideNavComponent implements OnInit, OnDestroy {
+export class MagicSideNavComponent implements OnInit, OnDestroy, AfterViewInit {
   private readonly _destroyRef = inject(DestroyRef);
   private readonly _sideNavConfigService = inject(MagicNavConfigService);
   private readonly _taskService = inject(TaskService);
   private readonly _layoutService = inject(LayoutService);
+  private _dragDropRegistry = inject(DragDropRegistry);
+  private _externalDragService = inject(ScheduleExternalDragService);
+  private _pointerUpSubscription: Subscription | null = null;
   // Use service's computed signal directly
   readonly config = this._sideNavConfigService.navConfig;
+  readonly WorkContextType = WorkContextType;
 
   activeWorkContextId = input<string | null>(null);
 
@@ -140,6 +151,11 @@ export class MagicSideNavComponent implements OnInit, OnDestroy {
       window.clearTimeout(this._animateTimeoutId);
       this._animateTimeoutId = null;
     }
+
+    if (this._pointerUpSubscription) {
+      this._pointerUpSubscription.unsubscribe();
+      this._pointerUpSubscription = null;
+    }
   }
 
   ngOnInit(): void {
@@ -168,6 +184,13 @@ export class MagicSideNavComponent implements OnInit, OnDestroy {
       this.isFullMode.set(this.config().fullModeByDefault);
       this.currentWidth.set(this.config().defaultWidth);
     }
+  }
+
+  ngAfterViewInit(): void {
+    // Listen for global pointer releases while a drag is active
+    this._pointerUpSubscription = this._dragDropRegistry.pointerUp.subscribe((event) => {
+      this._handlePointerUp(event);
+    });
   }
 
   onNavKeyDown(event: KeyboardEvent): void {
@@ -483,5 +506,64 @@ export class MagicSideNavComponent implements OnInit, OnDestroy {
     if (focusableElements.length > 0) {
       focusableElements[0].focus();
     }
+  }
+
+  private _handlePointerUp(event: MouseEvent | TouchEvent): void {
+    const draggedTask = this._externalDragService.activeTask();
+
+    // exclude subtasks and recurring tasks
+    if (!draggedTask || draggedTask.parentId || draggedTask.repeatCfgId) {
+      return;
+    }
+
+    const pointerPos = this._getPointerPosition(event);
+    if (!pointerPos) {
+      return;
+    }
+
+    const navItemElement = document
+      .elementFromPoint(pointerPos.x, pointerPos.y)
+      ?.closest('nav-item');
+    if (!navItemElement) {
+      return;
+    }
+
+    if (navItemElement.hasAttribute('data-project-id')) {
+      // Task is dropped on a project
+      const projectId = navItemElement.getAttribute('data-project-id');
+      Log.debug('Task dropped on Project', { draggedTask, projectId });
+      this._taskService.moveToProject(draggedTask, projectId!);
+    } else if (navItemElement.hasAttribute('data-tag-id')) {
+      // Task is dropped on a tag
+      const tagId = navItemElement.getAttribute('data-tag-id');
+      Log.debug('Task dropped on Tag', { draggedTask, tagId });
+
+      // Special case: "Today" tag means to schedule task for today
+      if (tagId === TODAY_TAG.id) {
+        this._taskService.addToToday(draggedTask);
+      } else {
+        if (!draggedTask.tagIds.includes(tagId!)) {
+          // tag not yet assigned to task, add it
+          this._taskService.updateTags(draggedTask, [...draggedTask.tagIds, tagId!]);
+        } else {
+          // tag already assigned to task, remove it
+          this._taskService.updateTags(
+            draggedTask,
+            draggedTask.tagIds.filter((t) => t !== tagId),
+          );
+        }
+      }
+    }
+  }
+
+  private _getPointerPosition(
+    event: MouseEvent | TouchEvent,
+  ): { x: number; y: number } | null {
+    if (!('touches' in event)) {
+      return { x: event.clientX, y: event.clientY };
+    }
+
+    const touch = event.touches[0] ?? event.changedTouches?.[0];
+    return touch ? { x: touch.clientX, y: touch.clientY } : null;
   }
 }
