+++ src/app/core-ui/magic-side-nav/nav-item/nav-item.component.ts
@@ -5,9 +5,6 @@ import {
   inject,
   input,
   output,
-  AfterViewInit,
-  OnDestroy,
-  ElementRef,
 } from '@angular/core';
 import { RouterLink, RouterModule } from '@angular/router';
 
@@ -20,7 +17,7 @@ import { DEFAULT_PROJECT_ICON } from '../../../features/project/project.const';
 import { WorkContextMenuComponent } from '../../work-context-menu/work-context-menu.component';
 import { FolderContextMenuComponent } from '../../folder-context-menu/folder-context-menu.component';
 import { ContextMenuComponent } from '../../../ui/context-menu/context-menu.component';
-import { CdkDragPlaceholder, DragDropRegistry } from '@angular/cdk/drag-drop';
+import { CdkDragPlaceholder } from '@angular/cdk/drag-drop';
 import { MatIconButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { MatMenuItem, MatMenuModule } from '@angular/material/menu';
@@ -30,11 +27,6 @@ import { Store } from '@ngrx/store';
 import { TranslatePipe } from '@ngx-translate/core';
 import { isSingleEmoji } from '../../../util/extract-first-emoji';
 import { MenuTreeKind } from '../../../features/menu-tree/store/menu-tree.model';
-import { Subscription } from 'rxjs';
-import { ScheduleExternalDragService } from '../../../features/schedule/schedule-week/schedule-external-drag.service';
-import { Log } from '../../../core/log';
-import { TODAY_TAG } from 'src/app/features/tag/tag.const';
-import { TaskService } from 'src/app/features/tasks/task.service';
 
 @Component({
   selector: 'nav-item',
@@ -67,13 +59,8 @@ import { TaskService } from 'src/app/features/tasks/task.service';
   },
   standalone: true,
 })
-export class NavItemComponent implements AfterViewInit, OnDestroy {
+export class NavItemComponent {
   private readonly _store = inject(Store);
-  private _dragDropRegistry = inject(DragDropRegistry);
-  private _externalDragService = inject(ScheduleExternalDragService);
-  private _pointerUpSubscription: Subscription | null = null;
-  private _elementRef = inject(ElementRef);
-  private _taskService = inject(TaskService);
 
   mode = input<'work' | 'folder' | 'row'>('work');
   variant = input<'default' | 'nav'>('default');
@@ -155,80 +142,4 @@ export class NavItemComponent implements AfterViewInit, OnDestroy {
     const iconValue = this.icon();
     return iconValue ? isSingleEmoji(iconValue) : false;
   });
-
-  ngAfterViewInit(): void {
-    // Listen for global pointer releases while a drag is active
-    this._pointerUpSubscription = this._dragDropRegistry.pointerUp.subscribe((event) => {
-      this._handlePointerUp(event);
-    });
-  }
-
-  ngOnDestroy(): void {
-    if (this._pointerUpSubscription) {
-      this._pointerUpSubscription.unsubscribe();
-      this._pointerUpSubscription = null;
-    }
-  }
-
-  private _handlePointerUp(event: MouseEvent | TouchEvent): void {
-    const draggedTask = this._externalDragService.activeTask();
-
-    // exclude subtasks and recurring tasks
-    if (!draggedTask || draggedTask.parentId || draggedTask.repeatCfgId) {
-      return;
-    }
-
-    if (this._isEventInsideElement(event)) {
-      if (this.type() === WorkContextType.PROJECT) {
-        // Task is dropped in a project
-        const projectId = this.workContext()?.id;
-        Log.debug('Task dropped on Project', { draggedTask, projectId });
-        this._taskService.moveToProject(draggedTask, projectId!);
-      } else if (this.type() === WorkContextType.TAG) {
-        // Task is dropped on a tag
-        const tagId = this.workContext()?.id;
-        Log.debug('Task dropped on Tag', { draggedTask, tagId });
-
-        // Special case: "Today" tag means to schedule task for today
-        if (tagId === TODAY_TAG.id) {
-          this._taskService.addToToday(draggedTask);
-        } else {
-          if (!draggedTask.tagIds.includes(tagId!)) {
-            // tag not yet assigned to task, add it
-            this._taskService.updateTags(draggedTask, [...draggedTask.tagIds, tagId!]);
-          } else {
-            // tag already assigned to task, remove it
-            this._taskService.updateTags(
-              draggedTask,
-              draggedTask.tagIds.filter((t) => t !== tagId),
-            );
-          }
-        }
-      }
-    }
-  }
-
-  private _getPointerPosition(
-    event: MouseEvent | TouchEvent,
-  ): { x: number; y: number } | null {
-    if (!('touches' in event)) {
-      return { x: event.clientX, y: event.clientY };
-    }
-
-    const touch = event.touches[0] ?? event.changedTouches?.[0];
-    return touch ? { x: touch.clientX, y: touch.clientY } : null;
-  }
-
-  private _isEventInsideElement(event: MouseEvent | TouchEvent): boolean {
-    const pointer = this._getPointerPosition(event);
-    const rect = this._elementRef.nativeElement.getBoundingClientRect();
-
-    return (
-      pointer != null &&
-      pointer.x >= rect.left &&
-      pointer.x <= rect.right &&
-      pointer.y >= rect.top &&
-      pointer.y <= rect.bottom
-    );
-  }
 }
